Add files via upload
XYZ company is a fast-growing company in Eastern Australia with more than 2 million customers globally. The company deals with the selling and buying of food items, which are basically operated from the headquarters. The company is intending to open a branch near the local village of Bonalbo. Thus, the company requires young IT graduates to design the network for the branch. The network is intended to operate separately from the HQ network.
Being a small network, the company has the following requirements during implementation;

a) One router and one switch to be used (all CISCO products).
b) 3 departments (Admin/IT, Finance/HR, and Customer service/Reception)
c) Each department is required to be in different VLANS.
d) Each department is required to have a wireless network for the users.
e) Host devices in the network are required to obtain an IPv4 address automatically.
f) Devices in all the departments are required to communicate with each other.

Assume the ISP gave out a base network of 192.168.1.0. You, as the young network engineer who has been hired, design and implement a network considering the above requirements.